fix tests imports
